refactor: organize directory structure
