+++ src/lib/types/image.d.ts
