+++ src/views/ProjectPage/components/project/ProjectCard.tsx
@@ -2,8 +2,8 @@ import styled from '@emotion/styled';
 import Image from 'next/image';
 import Link from 'next/link';
 import { useIsDesktop, useIsMobile, useIsTablet } from '@src/hooks/useDevice';
+import { GTM_CLASS } from '@src/lib/constants/gtmClass';
 import { LinkType, ProjectType } from '@src/lib/types/project';
-import { GTM_CLASS } from '@src/utils/const/gtmClass';
 import shortid from 'shortid';
 import { ReactComponent as AppStore } from '../../assets/app-store-40x40.svg';
 import { ReactComponent as Github } from '../../assets/github-40x40.svg';
